Source Code
A personal voice assistant created using Python is a remarkable application that seamlessly integrates the power of natural language processing and automation. Leveraging libraries such as Speech Recognition and pyttsx3, Python enables the development of a sophisticated voice-activated system. Users can interact with the assistant by speaking commands or asking questions, and the program interprets and processes the voice input.